% This file is part of the 'muste' package for R

\name{muste_ExpandFileName}
\alias{muste_ExpandFileName}
\title{Expand file name using Survo R paths}
\description{
  Expands file name containing Survo R paths 
}
\usage{
muste_ExpandFileName(path)
}
\arguments{
  \item{path}{input path}
}
\details{
  Gives expanded path as output.
}
\value{
  Expanded path.
}
\author{Reijo Sund}
%\examples{
%muste_ExpandFileName("<Survo>/TEST.SVO")
%}
\keyword{file}