\name{survo}
\alias{muste}
\alias{survo}
\alias{muste-package}
\title{Launch Survo editor}
\description{
  Initializes Survo R and opens the editorial interface window.
  For more information, type HELP? in the editor and activate by ESC-key.
  See also \href{http://www.survo.fi/muste/}{Survo R web pages} and
  \href{http://www.survo.fi/english/}{Survo web pages} for more information.
  Especially \href{http://www.survo.fi/tour/}{getting started tour},
      \href{http://www.survo.fi/books/}{Survo books},
      \href{http://www.survo.fi/demos/}{Survo demos} and
      \href{http://www.survo.fi/help/}{Survo help system}
  are likely to be useful for a newbie.
}
\usage{
  survo(sucro = "<empty>")
}
\arguments{
\item{sucro}{Path for start sucro.} 
}
